move PooledTxsRequestor to Eth namespace
