+++ src/Nethermind/Nethermind.Network/P2P/Subprotocols/Eth/IPooledTxsRequestor.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using Nethermind.Core.Crypto;
 using Nethermind.Network.P2P.Subprotocols.Eth.V65.Messages;
 
-namespace Nethermind.Network.P2P.Subprotocols.Eth.V65
+namespace Nethermind.Network.P2P.Subprotocols.Eth
 {
     public interface IPooledTxsRequestor
     {
